Display time precisely

+++ cpp/sphere.cpp
@@ -26,8 +26,8 @@ int main() {
     }
 
     clock_t end = clock();
-    const int time = int(double(end - start) / CLOCKS_PER_SEC);
-    printf("Process finished. (%d s)\n", time);
+    const double time = double(end - start) / CLOCKS_PER_SEC;
+    printf("Process finished. (%2.1f s)\n", time);
 
     return 0;
 }
